requirements.txt
vanna==0.7.9    
qdrant-client==1.15.1
google-generativeai==0.8.5
python-certifi-win32==1.6.1
python-dotenv==1.1.1
Flask==3.1.2
sentence-transformers==5.1.1
torch==2.8.0
transformers==4.57.0
tokenizers==0.22.1
safetensors==0.6.2
numpy==2.3.4
pandas==2.3.3
requests==2.32.5
tqdm==4.67.1
PyYAML==6.0.3
gspread==6.2.1
huggingface_hub==0.35.3